+++ engine/src/libCoreCommon/connection_to_master.h
@@ -15,6 +15,6 @@ namespace core
 
 		virtual void	onConnect(const std::string& szContext);
 		virtual void	onDisconnect();
-		virtual void	onDispatch(uint16_t nMsgType, const void* pData, uint16_t nSize);
+		virtual void	onDispatch(uint32_t nMsgType, const void* pData, uint16_t nSize);
 	};
 }
